+++ lib/models/userEditRadioModel.dart
@@ -1,0 +1,12 @@
+class RadioModel {
+  bool isSelected;
+  final int multiplaier;
+  final String text;
+  final String icon;
+  final String title;
+  final String subtitle;
+  final double padding;
+  final String subSubTitle;
+
+  RadioModel(this.isSelected, this.multiplaier, this.text, this.icon, this.title, this.subtitle,this.subSubTitle , this.padding,);
+}
