+++ lib/widgets/navigator/navigatorBody.dart
@@ -1,5 +1,3 @@
-import 'package:bubble_bottom_bar/bubble_bottom_bar.dart';
-import 'package:calory_calc/design/theme.dart';
 import 'package:ff_navigation_bar/ff_navigation_bar.dart';
 import 'package:flutter/material.dart';
