+++ pubspec.yaml
@@ -25,7 +25,6 @@ dependencies:
   flutter_native_admob: ^2.1.0
   carousel_slider: ^2.1.0
   flutter_config: ^1.0.7
-  bubble_bottom_bar: ^1.2.0
   ff_navigation_bar: ^0.1.4
